// src/assets/charting_library/charting_library.min.d.ts
export interface ChartingLibraryWidgetOptions {
  container_id: string;
  datafeed: IBasicDataFeed | (IBasicDataFeed & IDatafeedQuotesApi);
  interval: ResolutionString;
  symbol: string;
  auto_save_delay?: number;
  autosize?: boolean;
  debug?: boolean;
  disabled_features?: string[];
  drawings_access?: AccessList;
  enabled_features?: string[];
  fullscreen?: boolean;
  height?: number;
  library_path?: string;
  locale: LanguageCode;
  numeric_formatting?: NumericFormattingParams;
  saved_data?: object;
  studies_access?: AccessList;
  study_count_limit?: number;
  symbol_search_request_delay?: number;
  timeframe?: string;
  timezone?: 'exchange' | Timezone;
  toolbar_bg?: string;
  width?: number;
  charts_storage_url?: string;
  charts_storage_api_version?: AvailableSaveloadVersions;
  client_id?: string;
  user_id?: string;
  load_last_chart?: boolean;
  studies_overrides?: StudyOverrides;
  customFormatters?: CustomFormatters;
  overrides?: Overrides;
  snapshot_url?: string;
  indicators_file_name?: string;
  preset?: 'mobile';
  time_frames?: TimeFrameItem[];
  custom_css_url?: string;
  favorites?: Favorites;
  save_load_adapter?: IExternalSaveLoadAdapter;
  loading_screen?: LoadingScreenOptions;
  settings_adapter?: ISettingsAdapter;
}

export interface TradingTerminalWidgetOptions extends ChartingLibraryWidgetOptions {
  brokerConfig?: SingleBrokerMetaInfo;
  restConfig?: RestBrokerMetaInfo;
  widgetbar?: WidgetBarParams;
  rss_news_feed?: RssNewsFeedParams;
  news_provider?: NewsProvider;

  brokerFactory?(host: IBrokerConnectionAdapterHost): IBrokerWithoutRealtime | IBrokerTerminal;
}
// ...
export interface IChartingLibraryWidget {
  onChartReady(callback: EmptyCallback): void;

  onGrayedObjectClicked(callback: (obj: GrayedObject) => void): void;

  onShortcut(shortCut: string, callback: EmptyCallback): void;

  subscribe<EventName extends keyof SubscribeEventsMap>(event: EventName, callback: SubscribeEventsMap[EventName]): void;

  unsubscribe<EventName extends keyof SubscribeEventsMap>(event: EventName, callback: SubscribeEventsMap[EventName]): void;

  chart(index?: number): IChartWidgetApi;

  setLanguage(lang: LanguageCode): void;

  setSymbol(symbol: string, interval: ResolutionString, callback: EmptyCallback): void;

  remove(): void;

  closePopupsAndDialogs(): void;

  selectLineTool(linetool: SupportedLineTools): void;

  selectedLineTool(): SupportedLineTools;

  save(callback: (state: object) => void): void;

  load(state: object): void;

  getSavedCharts(callback: (chartRecords: SaveLoadChartRecord[]) => void): void;

  loadChartFromServer(chartRecord: SaveLoadChartRecord): void;

  saveChartToServer(onComplete?: EmptyCallback, onFail?: EmptyCallback, saveAsSnapshot?: false, options?: SaveChartToServerOptions): void;

  removeChartFromServer(chartId: string, onCompleteCallback: EmptyCallback): void;

  onContextMenu(callback: (unixTime: number, price: number) => ContextMenuItem[]): void;

  createButton(options?: CreateButtonOptions): JQuery;

  showNoticeDialog(params: DialogParams<() => void>): void;

  showConfirmDialog(params: DialogParams<(confirmed: boolean) => void>): void;

  showLoadChartDialog(): void;

  showSaveAsChartDialog(): void;

  symbolInterval(): SymbolIntervalResult;

  mainSeriesPriceFormatter(): IFormatter;

  getIntervals(): string[];

  getStudiesList(): string[];

  addCustomCSSFile(url: string): void;

  applyOverrides<TOverrides extends StudyOverrides>(overrides: TOverrides): void;

  applyStudiesOverrides(overrides: object): void;

  watchList(): WatchListApi;

  activeChart(): IChartWidgetApi;

  chartsCount(): number;

  layout(): LayoutType;

  setLayout(layout: LayoutType): void;
}

export interface ChartingLibraryWidgetConstructor {
  new (options: ChartingLibraryWidgetOptions | TradingTerminalWidgetOptions): IChartingLibraryWidget;
}

export declare function version(): string;

export declare function onready(callback: () => void): void;

export declare const widget: ChartingLibraryWidgetConstructor;